+++ drivers/net/wireless/iwlwifi/iwl4965-base.c
@@ -3119,7 +3119,7 @@ static int iwl4965_tx_status_reply_tx(struct iwl_priv *priv,
 		agg->rate_n_flags = le32_to_cpu(tx_resp->rate_n_flags);
 		IWL_DEBUG_TX_REPLY("Frames %d start_idx=%d bitmap=0x%llx\n",
 				   agg->frame_count, agg->start_idx,
-				   agg->bitmap);
+				   (unsigned long long)agg->bitmap);
 
 		if (bitmap)
 			agg->wait_for_ba = 1;
